+++ clients/vscode/media/canvas/chunk-chunk-55IACEB6.js
@@ -1,1 +1,1 @@
-import{_ as a}from"./chunk-mermaid.core.js";import{a2 as o}from"./index.js";var d=a((t,e)=>{let n;return e==="sandbox"&&(n=o("#i"+t)),(e==="sandbox"?o(n.nodes()[0].contentDocument.body):o("body")).select(`[id="${t}"]`)},"getDiagramElement");export{d as g};
+import{_ as a}from"./chunk-mermaid.core.js";import{a3 as o}from"./index.js";var d=a((t,e)=>{let n;return e==="sandbox"&&(n=o("#i"+t)),(e==="sandbox"?o(n.nodes()[0].contentDocument.body):o("body")).select(`[id="${t}"]`)},"getDiagramElement");export{d as g};